Add back OUT_DIR option to setup.sh
Since we're not using the WRF-setups directory as a generator for new runs
(i.e., we've abandoned the tagging system), each new WRF run will be generated
in a specified directory, following Pat H's original paradigm.

Also, added a check for existing OUT_DIR so that previous case setups will not
be overwritten by the setup script.