New 25>5>1 km setup for ERA5 w/MYJ on 5/15 to 5/17
Templates for the new 25 km to 1 km grid spacing setup using
MYJ PBL. The wrf_setup script calls the ERA5 reanalysis that
will need to be downloaded manually. Period of interest is
5/16 01Z to 05Z, thus, this simulation goes from 5/14 12Z to
5/17/ 00Z with the first 12 hours being spinup. Simulation
length can be shortened to end at 5/16 05Z if desired (be
sure to modify the template file before running the setup
script).